minor.
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/branches/v2.3_dev@11770 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ phar_installer/app/wizard/class.wizard_step7.php
@@ -154,6 +154,7 @@ class wizard_step7 extends \cms_autoinstaller\wizard_step
                                     $this->error('problem removing file '.$fn.'/index.html');
                                 }
                                 else {
+				    if( is_file($fn.'/moduleinfo.ini') ) $res = unlink($fn.'/moduleinfo.ini');
                                     $res = @rmdir($fn);
                                     if( !$res ) {
                                         $this->error('problem removing directory: '.$fn);
